pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cukereuse-subscenarios"
version = "0.1.0"
description = "Paraphrase-robust subsequence mining and three-scope refactoring-candidate ranking for Cucumber/Gherkin (paper 3 reproduction package)."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10,<3.13"
authors = [
  { name = "Ali Hassaan Mughal" },
  { name = "Noor Fatima" },
  { name = "Muhammad Bilal" },
]
keywords = ["cucumber", "gherkin", "bdd", "subsequence-mining", "test-refactoring", "clone-detection"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "numpy>=1.26,<2.0",
  "pandas>=2.2,<3.0",
  "pyarrow>=16,<17",
  "scikit-learn>=1.4,<2.0",
  "scipy>=1.11,<2.0",
  "sentence-transformers>=3,<4",
  "torch>=2.1,<3.0",
  "umap-learn>=0.5,<0.6",
  "hdbscan>=0.8,<0.9",
  "xgboost>=2.0,<3.0",
  "matplotlib>=3.8,<4.0",
  "tqdm>=4.66,<5.0",
  "requests>=2.31,<3.0",
]

[project.urls]
Repository = "https://github.com/amughalbscs16/cukereuse_subscenarios_release"
Issues    = "https://github.com/amughalbscs16/cukereuse_subscenarios_release/issues"
PaperOne  = "https://arxiv.org/abs/2604.20462"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["scripts"]
